pyproject.toml
[tool.poetry]
name = "project-1-analisis-of-banking-appkications"
version = "0.1.0"
description = ""
authors = ["Aleksandr Nikitin <NAI_surgut@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"


[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '.venv'

[tool.black]
line-length = 119
exclude = '.venv'

[tool.isort]

[flake8]
max-line-length = 119
exclude = '.git,__pycache__'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
